+++ backend/seeders/seed-characters.js
@@ -1,0 +1,2 @@
+// SEED CHARACTERS
+const db = require('../models')